+++ src/linked_list.js
@@ -1,3 +1,5 @@
+/* eslint-disable max-classes-per-file */
+
 /*!
  * This module defines the LinkedList class, which represents a doubly-linked list data structure.
  *