+++ test/health_checker.js
@@ -18,9 +18,9 @@ describe('HealthChecker', () => {
     };
 
     const writeTestCollection = {
-        insert: sinon.stub(),
+        insertOne: sinon.stub(),
         findOne: sinon.stub(),
-        remove: sinon.stub(),
+        deleteMany: sinon.stub(),
     };
 
     const mongoWrite = {
@@ -56,9 +56,9 @@ describe('HealthChecker', () => {
         sinon.assert.calledOnce(redis.set);
         sinon.assert.calledOnce(redis.get);
         sinon.assert.calledOnce(mongoRead.listCollections);
-        sinon.assert.calledOnce(writeTestCollection.insert);
+        sinon.assert.calledOnce(writeTestCollection.insertOne);
         sinon.assert.calledOnce(writeTestCollection.findOne);
-        sinon.assert.calledOnce(writeTestCollection.remove);
+        sinon.assert.calledOnce(writeTestCollection.deleteMany);
     });
 
     it('should fail when some of the checks returns something wrong.', async () => {
@@ -73,9 +73,9 @@ describe('HealthChecker', () => {
         sinon.assert.calledOnce(redis.set);
         sinon.assert.calledOnce(redis.get);
         sinon.assert.calledOnce(mongoRead.listCollections);
-        sinon.assert.calledOnce(writeTestCollection.insert);
+        sinon.assert.calledOnce(writeTestCollection.insertOne);
         sinon.assert.calledOnce(writeTestCollection.findOne);
-        sinon.assert.calledOnce(writeTestCollection.remove);
+        sinon.assert.calledOnce(writeTestCollection.deleteMany);
     });
 
     it('should fail when some of the checks returns something wrong.', async () => {
@@ -90,13 +90,13 @@ describe('HealthChecker', () => {
         sinon.assert.calledOnce(redis.set);
         sinon.assert.calledOnce(redis.get);
         sinon.assert.calledOnce(mongoRead.listCollections);
-        sinon.assert.calledOnce(writeTestCollection.insert);
+        sinon.assert.calledOnce(writeTestCollection.insertOne);
         sinon.assert.calledOnce(writeTestCollection.findOne);
-        sinon.assert.calledOnce(writeTestCollection.remove);
+        sinon.assert.calledOnce(writeTestCollection.deleteMany);
     });
 
     it('should fail when some of the checks timeouts', async () => {
-        redis.get.onFirstCall().returns(new Promise(resolve => setTimeout(resolve, 200))); // Timeouts.
+        redis.get.onFirstCall().returns(new Promise((resolve) => setTimeout(resolve, 200))); // Timeouts.
 
         await expect(healthChecker.ensureIsHealthy()).to.be.rejectedWith('Health check test "REDIS" failed with an error: Check has timed-out');
 
@@ -128,9 +128,9 @@ describe('HealthChecker', () => {
         sinon.assert.calledOnce(redis.set);
         sinon.assert.calledOnce(redis.get);
         sinon.assert.calledOnce(mongoRead.listCollections);
-        sinon.assert.calledOnce(writeTestCollection.insert);
+        sinon.assert.calledOnce(writeTestCollection.insertOne);
         sinon.assert.calledOnce(writeTestCollection.findOne);
-        sinon.assert.calledOnce(writeTestCollection.remove);
+        sinon.assert.calledOnce(writeTestCollection.deleteMany);
     });
 
     it('should fail when mongo cannot read', async () => {
@@ -143,9 +143,9 @@ describe('HealthChecker', () => {
         sinon.assert.calledOnce(redis.set);
         sinon.assert.calledOnce(redis.get);
         sinon.assert.calledOnce(mongoRead.listCollections);
-        sinon.assert.calledOnce(writeTestCollection.insert);
+        sinon.assert.calledOnce(writeTestCollection.insertOne);
         sinon.assert.calledOnce(writeTestCollection.findOne);
-        sinon.assert.calledOnce(writeTestCollection.remove);
+        sinon.assert.calledOnce(writeTestCollection.deleteMany);
     });
 
     it('should correctly validate checks', async () => {